fin jeudi

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: prigaudi <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 14:26:26 by prigaudi          #+#    #+#             */
-/*   Updated: 2024/11/06 14:52:34 by prigaudi         ###   ########.fr       */
+/*   Updated: 2024/11/14 15:08:06 by prigaudi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 size_t	ft_strlcat(char *dest, const char *src, size_t size)
 {
-	int	i;
-	int	destlen;
+	unsigned long	i;
+	int				destlen;
 
 	destlen = ft_strlen(dest);
 	i = 0;
@@ -28,12 +28,3 @@ size_t	ft_strlcat(char *dest, const char *src, size_t size)
 	dest[destlen + i] = '\0';
 	return (ft_strlen(dest) + ft_strlen(src));
 }
-
-// int	main(void)
-// {
-// 	char dest[100] = "coucou";
-// 	char *src = "salut";
-// 	int size = 4;
-// 	printf("resultat = %zu \n", ft_strlcat(dest, src, size));
-// 	printf("resultat2 = %s \n", dest);
-// }
